controller/certificate: move cert generation to cert controller
Relocate default wildcard certificate generation to the certificate controller to more
clearly decouple certificate management from ingress reconcilication.

Change the certificate controller event handling so that reconciliation is
triggered by clusteringress events. This makes the reconciliation code in the
controller easier to reason about and means we only do cert management if
there's a clusteringress around. It also means we can easily react to
clusteringresses being added/deleted.
