+++ pkg/operator/operator.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"github.com/openshift/cluster-ingress-operator/pkg/apis"
-	ingressv1alpha1 "github.com/openshift/cluster-ingress-operator/pkg/apis/ingress/v1alpha1"
 	"github.com/openshift/cluster-ingress-operator/pkg/dns"
 	logf "github.com/openshift/cluster-ingress-operator/pkg/log"
 	"github.com/openshift/cluster-ingress-operator/pkg/manifests"
@@ -30,7 +29,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -71,8 +69,6 @@ type Operator struct {
 
 	manager manager.Manager
 	caches  []cache.Cache
-	// certEvents is used to trigger the certificate controller.
-	certEvents chan event.GenericEvent
 }
 
 // New creates (but does not start) a new operator from configuration.
@@ -149,15 +145,14 @@ func New(config operatorconfig.Config, dnsManager dns.Manager, kubeConfig *rest.
 		})
 	}
 
-	certEvents := make(chan event.GenericEvent)
-	if _, err := certcontroller.New(operatorManager, kubeClient, config.Namespace, certEvents); err != nil {
+	// Set up the certificate controller
+	if _, err := certcontroller.New(operatorManager, kubeClient, config.Namespace); err != nil {
 		return nil, fmt.Errorf("failed to create cacert controller: %v", err)
 	}
 
 	return &Operator{
-		manager:    operatorManager,
-		caches:     []cache.Cache{operandCache},
-		certEvents: certEvents,
+		manager: operatorManager,
+		caches:  []cache.Cache{operandCache},
 
 		// TODO: These are only needed for the default cluster ingress stuff, which
 		// should be refactored away.
@@ -190,12 +185,6 @@ func (o *Operator) Start(stop <-chan struct{}) error {
 		}
 		log.Info("cache synced")
 	}
-
-	// Kick off the certificate controller
-	go func() {
-		ci := ingressv1alpha1.ClusterIngress{}
-		o.certEvents <- event.GenericEvent{Meta: ci.GetObjectMeta()}
-	}()
 
 	// Secondary caches are all synced, so start the manager.
 	go func() {
